Staging

+++ app/Http/Controllers/StockController.php
@@ -3,6 +3,10 @@
 namespace App\Http\Controllers;
 
 use App\Models\Stock;
+use App\Exports\StockExport;
+use Barryvdh\DomPDF\Facade\Pdf;
+use Maatwebsite\Excel\Facades\Excel;
+use Illuminate\Http\Request;
 
 class StockController extends Controller
 {
@@ -15,6 +19,18 @@ class StockController extends Controller
         );
     }
 
+    public function export(Request $request)
+    {
+        if ($request->format == 1) {
+            return Excel::download(new StockExport, 'Data-Stock.xlsx');
+        }
+
+        $stocks = Stock::all();
+        $pdf = Pdf::loadView('pages.exports.stock', compact('stocks'))
+            ->setPaper('a4', 'landscape');
+        return $pdf->download('Data-Stock.pdf');
+    }
+
     public function store()
     {
         // validate the request
@@ -22,21 +38,23 @@ class StockController extends Controller
             'stock_name' => 'required',
             'stock_quantity' => 'required|numeric',
             "stock_satuan" => "required",
-            "stock_description" => "required",
-            'stock_photo' => 'required|image|mimes:jpeg,png,jpg|max:1024',
+            'stock_photo' => 'image|mimes:jpeg,png,jpg|max:1024',
         ]);
 
         // get file
         $file = request()->file('stock_photo');
-        $file_name = md5($file->getClientOriginalName() . time()) . '.' . $file->getClientOriginalExtension();
-        $file->move('storage/stocks', $file_name);
+
+        if ($file) {
+            $file_name = md5($file->getClientOriginalName() . time()) . '.' . $file->getClientOriginalExtension();
+            $file->move('storage/stocks', $file_name);
+        }
 
         Stock::create([
             'stock_name' => request('stock_name'),
             'stock_quantity' => request('stock_quantity'),
             'stock_satuan' => request('stock_satuan'),
-            'stock_description' => request('stock_description'),
-            'stock_photo' => $file_name,
+            'stock_description' => request('stock_description') ?? "",
+            'stock_photo' => $file_name ?? "",
         ]);
         return back()->with('stock.success', 'Stock berhasil ditambahkan.');
     }
@@ -57,7 +75,9 @@ class StockController extends Controller
         if (request()->hasFile('stock_photo')) {
             // unlink old file
             $old = Stock::where("stock_id", $id)->first();
-            unlink(public_path('storage/stocks/' . $old->stock_photo));
+            if ($old->stock_photo && file_exists(public_path('storage/stocks/' . $old->stock_photo))) {
+                unlink(public_path('storage/stocks/' . $old->stock_photo));
+            }
 
             $file = request()->file('stock_photo');
             $file_name = md5($file->getClientOriginalName() . time()) . '.' . $file->getClientOriginalExtension();
@@ -89,7 +109,10 @@ class StockController extends Controller
 
             // unlink the file
             $stock = Stock::where("stock_id", $id)->first();
-            unlink(public_path('storage/stocks/' . $stock->stock_photo));
+
+            if ($stock->stock_photo && file_exists(public_path('storage/stocks/' . $stock->stock_photo))) {
+                unlink(public_path('storage/stocks/' . $stock->stock_photo));
+            }
 
             Stock::where("stock_id", $id)->delete();
 
@@ -100,7 +123,7 @@ class StockController extends Controller
         } catch (\Throwable $th) {
             return response()->json([
                 "status" => false,
-                "message" => "Stock gagal dihapus.",
+                "message" => $th->getMessage() ?? "Stock gagal dihapus",
             ])->setStatusCode(500);
         }
     }
